mobx: Refactor the AlertRule into its own file and add a helper-file
